requirements.txt
aiohttp
https://github.com/New-dev0/Telethon/archive/diamond.zip
html_telegraph_poster
lxml[html_clean]
python-decouple
psutil
pymediainfo
tgcrypto
nvidia-ml-py3
gpustat
pynvml
